Removing hashes in Asset urls.
By replacing # with / in asset urls, everything works! There is no
need for url rewrites in css files as all files stay relative to each
other.

In AssetsController, the module prefix is found by looking at the
first segment of the asset before the first slash. To work with
assets not part of a module, put them in public - as they are files
the request will never reach AssetsController anyway.