Refactoring Route to use http-foundation Request.
Route#test() should now be called with a Request instance. From there
Route can work out everything about the request without relying on any
globals. YEAH.

Modifying Dispatcher with match and matchRequest methods. matchRequest
is the default for routing requests. match could used for testing a
single url, as it mocks a Request instance to be used for testing. This
could be used for a future router:test command.