Cleaning up Route.
Route no longer relies on the RequestHelper class, as it now examines
the entire url instead of removing the extension. Instead a magic
_format variable is added to the url regex, so format testing can be
done automatically with calls to format().

Adding argsRegex() to decide on what characters are allowed in variables
- this remove ARGS_SINGLE, as a regex of .+ can be used. Now that
variables by default DON'T include a dot, callArgsRegex('[^/]+') to
allow for arguments with a dot. This will remove the auto format, however.

Also improving usage of 'auto args' - calling autoArgs() will replace
:args with a populated array of auto arguments.