Removing RequestHelper class.
So begins the trimming down of this library! I plan to strip out forms
and security into reusable modules, and essentially freeze any further
progress on this framework. It's been fun and a great learning
experience. Let's finish this thing - clean up the code, add some docs
and bump to version 1 - the first and last major release. I want this
release to be stable and usable, with only bugfixes following this release.